fix frost vinette render
