+++ src/main/java/com/github/thedeathlycow/frostiful/mixins/client/gui/InGameHudMixin.java
@@ -1,7 +1,10 @@
 package com.github.thedeathlycow.frostiful.mixins.client.gui;
 
+import com.github.thedeathlycow.frostiful.client.gui.FrostOverlayRenderer;
 import com.github.thedeathlycow.frostiful.client.gui.RootedOverlayRenderer;
 import com.github.thedeathlycow.frostiful.entity.RootedEntity;
+import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
+import com.llamalad7.mixinextras.injector.v2.WrapWithCondition;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.gui.DrawContext;
 import net.minecraft.client.gui.hud.InGameHud;
@@ -12,6 +15,8 @@ import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.ModifyArg;
+import org.spongepowered.asm.mixin.injection.Slice;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(InGameHud.class)
@@ -25,6 +30,28 @@ public abstract class InGameHudMixin {
     protected abstract void renderOverlay(DrawContext context, Identifier texture, float opacity);
 
 
+    @WrapWithCondition(
+            method = "renderMiscOverlays",
+            slice = @Slice(
+                    from = @At(
+                            value = "INVOKE",
+                            target = "Lnet/minecraft/client/network/ClientPlayerEntity;getFrozenTicks()I"
+                    )
+            ),
+            at = @At(
+                    value = "INVOKE",
+                    target = "Lnet/minecraft/client/gui/hud/InGameHud;renderOverlay(Lnet/minecraft/client/gui/DrawContext;Lnet/minecraft/util/Identifier;F)V",
+                    ordinal = 0
+            )
+    )
+    private boolean blockVanillaFrozenOverlayRender(
+            InGameHud instance,
+            DrawContext context, Identifier texture, float opacity
+    ) {
+        return false;
+    }
+
+
     @Inject(
             method = "renderMiscOverlays",
             at = @At("TAIL")
@@ -34,6 +61,11 @@ public abstract class InGameHudMixin {
             RootedOverlayRenderer.render(
                     (RootedEntity) this.client.player,
                     context, tickCounter,
+                    this::renderOverlay
+            );
+            FrostOverlayRenderer.renderFrostOverlay(
+                    context,
+                    this.client.player,
                     this::renderOverlay
             );
         }
